2412120051      - #288 Updated Analog calcs to cast to double 2412120016      - Changed ADS1115 to 8 SPS                 - #288 removed struct from ADC calc                 - Renames buildSSEJsonData to buildSSEJsonData

+++ ESP32/DIY-Flow-Bench/structs.h
@@ -278,7 +278,7 @@ struct SensorData {
   double TempDegC = 0.0;
   double TempDegF = 0.0;
   double RelH = 0.0;
-  long BaroPA = 0;
+  double BaroPA = 0;
   double BaroKPA = 0.0;
   double BaroHPA = 0.0;
   double PRefKPA = 0.0;
